added optional command auravalue=#
forces Gnosis to display the aura value of the buff/debuff instead of the
duration, thereby converting the timer bar to a value based bar
(e.g. allowing mcnt and msize commands)

Gnosis does not know the maximum value for any Aura, so the user has to pass this maximum value
as a parameter (e.g. auravalue=5)